Beta-Navigation Environment Support
This is a series of changes to support a separate "feature branch" environment for
beta navigation-enabled versions of the UI.  This is accomplished by forcing domain recognition to obey
a specific prioritization of pattern matching per location type, from
production (first) to local development (last).  Thus more generic PR
branches will not be recognized before the feature branch.

This also includes a restructuring of the data structures used to
recognize domains and their associated context.  Previously, this was
accomplished using a brute-force array of `RegExp.test` calls.  This
more optimized implementation uses branching by keyword and dynamic
RegExp compilation -- since `String.indexOf` is dramatically faster than
even the simplest regular expression, and most regular expressions will
never need to be compiled in the first place, this increases the overall
performance of domain recognition by a factor of approximately 25
(in local tests, bringing the average down from about 0.65ms to
0.023ms).

Also ups the ante on testing coverage.