Menu Item - Publicity Inheritance
Added logic to the routing types to allow menu items or trees of menu
items (inheriting in a descending fashion) to be marked as public or
private.

If the route parameter `anonymous` is set to `"true"`, items whose
publicity is evaluated to `false` will be hidden.

If the route parameter `anonymous` is set to `"false"`, items whose
publicity is evaluated to `true` will be hidden.

`null` is ignored.